Component rework. Split types, use interface
