+++ esphome/components/micronova/sensor/micronova_sensor.h
@@ -1,0 +1,25 @@
+#pragma once
+
+#include "esphome/components/micronova/micronova.h"
+#include "esphome/components/sensor/sensor.h"
+
+namespace esphome {
+namespace micronova {
+
+class MicroNovaSensor : public sensor::Sensor, public MicroNovaSensorListener {
+ public:
+  MicroNovaSensor(MicroNova *m) : MicroNovaSensorListener(m) {}
+  void dump_config() override { LOG_SENSOR("", "Micronova sensor", this); }
+  void publish_val(int new_raw_value) override;
+  void set_stove_switch_state(bool v) {}
+  bool get_stove_switch_state() { return false; }
+
+  void set_fan_speed_offset(uint8_t f) { fan_speed_offset_ = f; }
+  uint8_t get_set_fan_speed_offset() { return fan_speed_offset_; }
+
+ protected:
+  int fan_speed_offset_ = 0;
+};
+
+}  // namespace micronova
+}  // namespace esphome
