+++ esphome/components/micronova/text_sensor/micronova_text_sensor.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "esphome/components/micronova/micronova.h"
+#include "esphome/components/text_sensor/text_sensor.h"
+
+namespace esphome {
+namespace micronova {
+
+class MicroNovaTextSensor : public text_sensor::TextSensor, public MicroNovaSensorListener {
+ public:
+  MicroNovaTextSensor(MicroNova *m) : MicroNovaSensorListener(m) {}
+  void dump_config() override { LOG_TEXT_SENSOR("", "Micronova text sensor", this); }
+  void publish_val(int new_raw_value) override;
+  void set_stove_switch_state(bool v) {}
+  bool get_stove_switch_state() { return false; }
+};
+
+}  // namespace micronova
+}  // namespace esphome
